requirements.txt
Django==3.0.4
django-environ
gunicorn
pyyaml
bs4
djangorestframework
psycopg2-binary
django-storages[google]
google-cloud-secret-manager==0.1.1